requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
snowflake-connector-python
snowflake-sqlalchemy
python-dotenv